+++ src/utils/redirect.ts
@@ -4,18 +4,25 @@ export const preventRedirects = (shouldPreventRedirect = true) => {
   preventRedirect = shouldPreventRedirect;
 };
 
-export const safeRedirect = (url: string, timeout = 0) => {
+export const safeRedirect = ({
+  shouldForcePageReload,
+  timeout = 0,
+  url
+}: {
+  shouldForcePageReload?: boolean;
+  timeout?: number;
+  url: string;
+}) => {
   if (!preventRedirect) {
     setTimeout(() => {
       if (!window) {
         return;
       }
 
-      // Navigate to callbackUrl without page refresh
-      // if we are in the same origin
+      // Navigate to callbackUrl without page refresh if we are in the same origin
       const isSameOriginRedirect = url?.startsWith('/');
 
-      if (isSameOriginRedirect) {
+      if (isSameOriginRedirect && !shouldForcePageReload) {
         return window.history.pushState('', '', url);
       }